+++ optimum/utils/__init__.py
@@ -12,7 +12,10 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import importlib.util
+import importlib
+from contextlib import contextmanager
 
+from packaging import version
 
 CONFIG_NAME = "config.json"
 
@@ -31,6 +34,29 @@ def is_pydantic_available():
 
 def is_accelerate_available():
     return _accelerate_available
+
+
+def is_pytorch_greater_112():
+    import torch
+
+    return version.parse(torch.__version__) >= version.parse("1.12.0")
+
+
+@contextmanager
+def check_if_pytorch_greater_112():
+    r"""
+    A context manager that does nothing except checking if the PyTorch version is greater than 1.12.0.
+    """
+    import torch
+
+    if not is_pytorch_greater_112():
+        raise ImportError(
+            f"Found an incompatible version of PyTorch. Found version {torch.__version__}, but only 1.12.0 and above are supported."
+        )
+    try:
+        yield
+    finally:
+        pass
 
 
 from .input_generators import (  # noqa
